+++ packages/nimble-components/src/text-area/tests/text-area-matrix.stories.ts
@@ -14,7 +14,9 @@ import {
     disabledStates,
     DisabledState,
     ReadOnlyState,
-    readOnlyStates
+    readOnlyStates,
+    ErrorState,
+    errorStates
 } from '../../utilities/tests/states';
 import { hiddenWrapper } from '../../utilities/tests/hidden';
 import { textCustomizationWrapper } from '../../utilities/tests/text-customization';
@@ -46,18 +48,21 @@ const component = (
     [readOnlyName, readonly]: ReadOnlyState,
     [disabledName, disabled]: DisabledState,
     [appearanceName, appearance]: AppearanceState,
-    [valueName, valueValue, placeholderValue]: ValueState
+    [valueName, valueValue, placeholderValue]: ValueState,
+    [errorStateName, isError, errorText]: ErrorState
 ): ViewTemplate => html`
     <${textAreaTag}
-        style="width: 250px; padding: 15px;"
+        style="width: 250px; margin: 15px;"
         ?disabled="${() => disabled}"
         appearance="${() => appearance}"
         value="${() => valueValue}"
         placeholder="${() => placeholderValue}"
         ?readonly="${() => readonly}"
+        error-visible="${() => isError}"
+        error-text="${() => errorText}"
     >
         ${() => disabledName} ${() => appearanceName} ${() => valueName}
-        ${() => readOnlyName}
+        ${() => readOnlyName} ${() => errorStateName}
     </${textAreaTag}>
 `;
 
@@ -66,7 +71,8 @@ export const textAreaThemeMatrix: StoryFn = createMatrixThemeStory(
         readOnlyStates,
         disabledStates,
         appearanceStates,
-        valueStates
+        valueStates,
+        errorStates
     ])
 );